+++ src/lib_smart_rollup_node/migrations/001_add_state_hash_and_status.sql
@@ -1,0 +1,7 @@
+--------------------------------------------------------
+-- SPDX-License-Identifier: MIT                       --
+-- Copyright (c) 2026 Functori <contact@functori.com> --
+--------------------------------------------------------
+
+ALTER TABLE l2_blocks ADD COLUMN state_hash VARCHAR(54);
+ALTER TABLE l2_blocks ADD COLUMN pvm_status BLOB;