+++ src/lib_smart_rollup_node/rollup_node_daemon.ml
@@ -179,7 +179,7 @@ let process_unseen_head ({node_ctxt; _} as state) ~catching_up ~predecessor
   (* Avoid storing and publishing commitments if the head is not final. *)
   (* Avoid triggering the pvm execution if this has been done before for
      this head. *)
-  let* _num_messages, num_ticks, initial_tick =
+  let* {num_ticks; initial_tick; state_hash; _} =
     Interpreter.process_head
       (module Plugin)
       node_ctxt
@@ -221,8 +221,28 @@ let process_unseen_head ({node_ctxt; _} as state) ~catching_up ~predecessor
         inbox_hash;
       }
   in
+  let* pvm_state = Context.PVMState.get ctxt in
+  let constants = (Reference.get node_ctxt.current_protocol).constants in
+  let* pvm_status =
+    Lwt.catch
+      (fun () ->
+        Plugin.Pvm.get_status node_ctxt ~constants pvm_state
+        |> Lwt_result.map Option.some)
+      (fun _ ->
+        (* Guard against partial implementations of status to string
+           conversions *)
+        Option.none_es)
+  in
   let l2_block =
-    Sc_rollup_block.{header; content = (); num_ticks; initial_tick}
+    Sc_rollup_block.
+      {
+        header;
+        content = ();
+        num_ticks;
+        initial_tick;
+        state_hash = Some state_hash;
+        pvm_status;
+      }
   in
   let* () = Node_context.save_l2_block node_ctxt l2_block in
   let* () =
@@ -814,7 +834,7 @@ module Internal_for_tests = struct
         ~predecessor
         messages
     in
-    let* _num_messages, num_ticks, initial_tick =
+    let* {num_ticks; initial_tick; state_hash; _} =
       Interpreter.process_head
         (module Plugin)
         node_ctxt
@@ -855,8 +875,19 @@ module Internal_for_tests = struct
           inbox_hash;
         }
     in
+    let* pvm_state = Context.PVMState.get ctxt in
+    let constants = (Reference.get node_ctxt.current_protocol).constants in
+    let* pvm_status = Plugin.Pvm.get_status node_ctxt ~constants pvm_state in
     let l2_block =
-      Sc_rollup_block.{header; content = (); num_ticks; initial_tick}
+      Sc_rollup_block.
+        {
+          header;
+          content = ();
+          num_ticks;
+          initial_tick;
+          state_hash = Some state_hash;
+          pvm_status = Some pvm_status;
+        }
     in
     let* () = Node_context.save_l2_block node_ctxt l2_block in
     let* () = Node_context.set_l2_head node_ctxt l2_block in
@@ -1150,7 +1181,7 @@ module Replay = struct
     let* (module Plugin) =
       Protocol_plugins.proto_plugin_for_level node_ctxt block.header.level
     in
-    let* _num_messages, num_ticks, initial_tick =
+    let* {num_ticks; initial_tick; _} =
       Interpreter.process_head
         (module Plugin)
         node_ctxt